refactor: enum 내에서 람다를 쓰는 것이 아닌 새로운 strategy 객체를 생성 (가독성 향상)
