+++ src/main/java/lotto/domain/AnswerLotto.java
@@ -3,6 +3,7 @@ package lotto.domain;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import lotto.domain.prizestrategy.LottoPrize;
 
 public class AnswerLotto {
