+++ src/main/java/lotto/domain/PrizeCountMap.java
@@ -4,6 +4,7 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
+import lotto.domain.prizestrategy.LottoPrize;
 
 public class PrizeCountMap {
