+++ src/main/java/lotto/domain/prizestrategy/FifthPrizeStrategy.java
@@ -1,0 +1,11 @@
+package lotto.domain.prizestrategy;
+
+import lotto.domain.Lotto;
+
+public class FifthPrizeStrategy implements PrizeStrategy {
+
+    @Override
+    public boolean isQualified(int matchNumberCount, boolean hasMagicNumber) {
+        return matchNumberCount == Lotto.LOTTO_NUMBER_LENGTH - 3;
+    }
+}
